# packages/contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: k8s-dashboard Observability API
  description: >-
    Kubernetes 운영 흐름에 특화된 화면 단위 조회 API입니다. 화면 하나 = 요청 하나(ADR 0002),
    Scope는 서버가 강제합니다(README §10). 부분 장애는 HTTP 에러가 아니라 Section 값입니다.
  version: 0.5.0
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/clusters/{clusterId}/resources:
    get:
      operationId: getResourceCatalog
      summary: "탐색 가능한 리소스 카탈로그 (platform.admin 전용, ADR 0018)"
      description: >-
        시작 시점과 주기적 discovery로 만든 불변 snapshot입니다. 요청 처리 중에는
        discovery도 Kubernetes LIST도 호출하지 않습니다. 각 항목의 state가
        ready·syncing·unsupported·forbidden·missing을 구분합니다.
      parameters:
        - $ref: "#/components/parameters/ClusterId"
      responses:
        "200":
          description: ResourceCatalogResponse
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceCatalogResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ResourceUnavailable"
  /api/v1/clusters/{clusterId}/resources/{group}/{version}/{resource}:
    get:
      operationId: listClusterResources
      summary: "allowlist 리소스 목록 (metadata 인덱스, keyset 페이지)"
      description: >-
        로컬 metadata informer 인덱스에서만 읽습니다. 정렬 키는 (namespace, name)
        하나이며 cursor는 opaque keyset입니다. offset 페이징은 없습니다 (ADR 0003).
        namespace 범위는 서버가 Scope와 교차시켜 강제 삽입합니다.
      parameters:
        - $ref: "#/components/parameters/ClusterId"
        - $ref: "#/components/parameters/ResourceGroup"
        - $ref: "#/components/parameters/ResourceVersion"
        - $ref: "#/components/parameters/ResourcePlural"
        - $ref: "#/components/parameters/Ns"
        - $ref: "#/components/parameters/ResourceNamePrefix"
        - $ref: "#/components/parameters/ResourceLabelSelector"
        - $ref: "#/components/parameters/ResourceLimit"
        - $ref: "#/components/parameters/ResourceCursor"
        - $ref: "#/components/parameters/ResourceSort"
        - $ref: "#/components/parameters/ResourceOrder"
      responses:
        "200":
          description: ResourceListResponse
          headers:
            X-Request-ID:
              $ref: "#/components/headers/XRequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "405":
          $ref: "#/components/responses/MethodNotAllowed"
        "500":
          $ref: "#/components/responses/InternalError"
        "502":
          $ref: "#/components/responses/ResourceUpstream"
        "503":
          $ref: "#/components/responses/ResourceUnavailable"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-k8s-dashboard
      description: Opaque same-origin HttpOnly browser-session handle; it is not an OIDC token and may outlive the server-side session.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        AUTH_MODE=oidc|mock일 때 필요한 Bearer JWT입니다. AUTH_MODE=none이면 정적 Scope로
        동작합니다. /healthz · /readyz · /version은 어느 모드에서든 인증이 필요 없습니다.
  headers:
    XRequestID:
      description: >-
        요청 상관관계 ID. 인바운드 X-Request-ID가 1..128자의 [A-Za-z0-9._:-]이면 재사용하고,
        아니면 서버가 128bit 난수 소문자 hex를 새로 만듭니다. 에러 본문 requestId와 같은 값입니다.
      schema:
        type: string
        pattern: "^[A-Za-z0-9._:-]{1,128}$"
...
  parameters:
...
    ClusterId:
      name: clusterId
      in: path
      required: true
      description: 클러스터 ID. Scope 밖이면 403입니다.
      schema:
        type: string
    Ns:
      name: ns
      in: query
      description: Scope와 교차해 적용할 namespace. 비우거나 all이면 Scope 전체입니다.
      schema:
        type: string
...
    ResourceGroup:
      name: group
      in: path
      required: true
      description: API group. core group은 "core"로 표기합니다.
      schema:
        type: string
        maxLength: 253
        pattern: "^[a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*$"
    ResourceVersion:
      name: version
      in: path
      required: true
      schema:
        type: string
        maxLength: 32
        pattern: "^v[0-9]+((alpha|beta)[0-9]+)?$"
    ResourcePlural:
      name: resource
      in: path
      required: true
      description: 복수형 리소스 이름. subresource는 대상이 아닙니다.
      schema:
        type: string
        maxLength: 63
        pattern: "^[a-z][a-z0-9]*$"
    ResourceNamePrefix:
      name: name
      in: query
      required: false
      description: 이름 prefix 필터.
      schema:
        type: string
        maxLength: 253
    ResourceLabelSelector:
      name: labelSelector
      in: query
      required: false
      description: Kubernetes label selector. 표현식은 최대 8개입니다.
      schema:
        type: string
        maxLength: 512
    ResourceLimit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    ResourceCursor:
      name: cursor
      in: query
      required: false
      description: >-
        opaque keyset cursor. 필터·정렬·Scope가 바뀐 cursor는 400입니다 —
        이어붙이면 중복이나 누락이 조용히 생기기 때문입니다.
      schema: {type: string, maxLength: 512, pattern: "^[A-Za-z0-9_-]+$"}
    ResourceSort:
      name: sort
      in: query
      required: false
      description: 정렬 키. 인덱스가 있는 (namespace, name) 하나뿐입니다.
      schema: {type: string, enum: [name], default: name}
    ResourceOrder:
      name: order
      in: query
      required: false
      schema: {type: string, enum: [asc, desc], default: asc}
...
  responses:
...
    BadRequest:
      description: 요청이 성립하지 않습니다 (예 invalid_range).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    Unauthorized:
      description: 인증이 필요합니다.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    Forbidden:
      description: 요청한 클러스터·namespace가 Scope 밖입니다. 부분 데이터도 내려가지 않습니다.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    NotFound:
      description: 등록되지 않은 경로입니다.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    MethodNotAllowed:
      description: 허용되지 않은 메서드입니다. Allow 헤더에 허용 메서드가 담깁니다.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
        Allow:
          description: 허용 메서드 (전 라우트 GET)
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    InternalError:
      description: 서버 내부 오류입니다. 내부 주소·질의·스택트레이스는 담기지 않습니다.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
    ResourceUnavailable:
      description: >-
        이 배포에는 Resource Explorer가 없거나(central·비활성) 캐시가 아직 동기화
        중입니다. 권한 판정 뒤에 나오는 상태이며, 코드가 resources_unavailable인지
        resource_syncing인지로 구분합니다.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
        Retry-After:
          description: 재시도까지 대기할 초(resource_syncing일 때).
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
    ResourceUpstream:
      description: >-
        클러스터가 이 리소스를 metadata 전용으로 제공하지 않거나(406) 서버 권한이
        없거나 객체가 응답 한도를 넘었습니다. 내부 주소·질의는 담기지 않습니다.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/XRequestID"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/APIError"
...
  schemas:
...
    APIError:
      type: object
      description: 화면 전체가 실패했을 때만 쓰는 최상위 에러. 섹션 단위 실패는 Section 값입니다.
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          enum:
            - unauthorized
            - forbidden
            - not_found
            - method_not_allowed
            - upstream_unavailable
            - invalid_range
            - bad_last_event_id
            - stream_capacity
            - internal
            - invalid_dashboard
            - invalid_revision
            - invalid_page
            - invalid_cursor
            - precondition_required
            - revision_conflict
            - draft_limit
            - approved_immutable
            - invalid_state
            - unsupported_media_type
            - body_too_large
            - dashboard_store_unavailable
            - bad_request
            - auth_unavailable
            - refresh_conflict
            - auth_rate_limited
            - session_unavailable
            - cluster_access_denied
            - namespace_access_denied
            - cluster_scope_required
            - resources_unavailable
            - resource_not_allowlisted
            - resource_not_served
            - resource_unsupported
            - resource_forbidden
            - resource_syncing
            - invalid_filter
            - detail_rate_limited
            - uid_mismatch
            - object_too_large
            - upstream_timeout
        message:
          type: string
        requestId:
          type: string
          description: 응답 헤더 X-Request-ID와 같은 값
...
    ResourceState:
      type: string
      description: allowlist 한 항목의 상태. "0건"과 "권한 없음"과 "미지원"을 구분합니다.
      enum: [ready, syncing, unsupported, forbidden, missing]
    ResourceDescriptor:
      type: object
      additionalProperties: false
      required: [group, version, resource, kind, namespaced, verbs, state, count]
      properties:
        group: {type: string, minLength: 1, description: 'core group은 "core"로 표기합니다.'}
        version: {type: string, minLength: 1}
        resource: {type: string, minLength: 1}
        kind: {type: string}
        namespaced: {type: boolean}
        verbs:
          type: array
          uniqueItems: true
          items: {type: string, minLength: 1}
        preferredVersion: {type: string}
        state: {$ref: "#/components/schemas/ResourceState"}
        reason: {type: string}
        count: {type: integer, minimum: 0}
    ResourceCatalogResponse:
      type: object
      additionalProperties: false
      required: [clusterId, generatedAt, degraded, items]
      properties:
        clusterId: {type: string, minLength: 1}
        generatedAt: {type: string, format: date-time}
        refreshedAt: {type: string, format: date-time}
        degraded: {type: boolean}
        reason: {type: string}
        items:
          type: array
          maxItems: 64
          items: {$ref: "#/components/schemas/ResourceDescriptor"}
    ResourceListItem:
      type: object
      additionalProperties: false
      required: [name, uid]
      properties:
        namespace: {type: string}
        name: {type: string, minLength: 1}
        uid: {type: string, minLength: 1}
        createdAt: {type: string, format: date-time}
    ResourceListResponse:
      type: object
      additionalProperties: false
      description: >-
        유계 keyset 페이지입니다. 응답 본문은 1MiB, 행은 200개를 넘지 않습니다.
      x-max-response-bytes: 1048576
      required: [clusterId, group, version, resource, kind, namespaced, generatedAt, appliedScope, items, truncated, total]
      properties:
        clusterId: {type: string, minLength: 1}
        group: {type: string, minLength: 1}
        version: {type: string, minLength: 1}
        resource: {type: string, minLength: 1}
        kind: {type: string}
        namespaced: {type: boolean}
        generatedAt: {type: string, format: date-time}
        observedAt: {type: string, format: date-time}
        appliedScope:
          type: object
          additionalProperties: false
          required: [clusterId, namespaces]
          properties:
            clusterId: {type: string, minLength: 1}
            namespaces:
              oneOf:
                - {type: string, const: all}
                - type: array
                  uniqueItems: true
                  items: {type: string, minLength: 1}
        items:
          type: array
          maxItems: 200
          items: {$ref: "#/components/schemas/ResourceListItem"}
        nextCursor: {type: string, maxLength: 512, pattern: "^[A-Za-z0-9_-]+$"}
        truncated: {type: boolean}
        total: {type: integer, minimum: 0}
...